switched times matcher to use NSInteger. Added TODO for variable numbers of args. -James

+++ Stubble/SBLTimesMatcher.h
@@ -3,11 +3,11 @@
 
 @interface SBLTimesMatcher : NSObject
 
-@property (nonatomic, readonly) int atLeast;
-@property (nonatomic, readonly) int atMost;
+@property (nonatomic, readonly) NSInteger atLeast;
+@property (nonatomic, readonly) NSInteger atMost;
 
 + (instancetype)never;
-+ (instancetype)exactly:(int)times;
-+ (instancetype)atLeast:(int)minTimes;
++ (instancetype)exactly:(NSInteger)times;
++ (instancetype)atLeast:(NSInteger)minTimes;
 
 @end
